Auto-fix Clippy warnings: 856 → 37 (95.7% reduction)
Applied cargo clippy --fix --allow-dirty to automatically fix safe warnings:
- 95 src files: format strings, bool comparisons, useless conversions
- 123 test files: assert_eq!(x, true) → assert!(x), format strings
- 46 bench files: format strings, single element loops

Results:
- Total warnings: 856 → 37 (95.7% reduction)
- All 3,681 tests still passing
- Remaining 37 warnings (by Pareto):
  * 16 needless_range_loop (43%) - for loops that could use iterators
  * 11 new_ret_no_self (30%) - factory methods named 'new'
  * 4 type_complexity (11%) - complex type signatures
  * 2 overly_complex_bool_expr (5%) - tautologies in tests
  * 2 ptr_arg (5%) - &Vec → &[_]
  * 2 empty_line_after_doc_comments (5%)

Changes:
- Updated analyses/clippy.txt with new output
- Modified 265 files (src, tests, benches)
- No manual changes - all auto-fixed by Clippy

+++ apas-ai/benches/Chap41/BenchArraySetStEph.rs
@@ -158,9 +158,9 @@ fn bench_array_set_from_seq(c: &mut Criterion) {
         // Create sequence with some duplicates
         let mut vec_data = Vec::new();
         for i in 0..*size {
-            vec_data.push(i as i32);
+            vec_data.push(i);
             if i % 3 == 0 {
-                vec_data.push(i as i32); // Add duplicate
+                vec_data.push(i); // Add duplicate
             }
         }
         let seq = ArraySeqStEphS::from_vec(vec_data);
